backend/pyproject.toml
[tool.poetry]
name = "career-predict-api"
version = "0.1.0"
description = "Adaptive career recommendation backend"
authors = ["BLACKBOXAI <ai@blackbox.ai>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
numpy = "^2.1.1"
pydantic = "^2.9.2"
python-multipart = "^0.0.9"
python-dotenv = "^1.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
